Reorganize convert (#423)
* initial organization

* moved all convert files

* added docs

* fix

* fixed errors

* format: isort

* reorganized tests
