+++ docs/source/api/convert/xgi.convert.bipartite_edges.rst
@@ -1,0 +1,11 @@
+xgi.convert.bipartite_edges
+===========================
+
+.. currentmodule:: xgi.convert.bipartite_edges
+
+.. automodule:: xgi.convert.bipartite_edges
+   
+   .. rubric:: Functions
+   
+   .. autofunction:: from_bipartite_edgelist
+   .. autofunction:: to_bipartite_edgelist
