+++ docs/source/api/convert/xgi.convert.graph.rst
@@ -1,0 +1,10 @@
+xgi.convert.graph
+=================
+
+.. currentmodule:: xgi.convert.graph
+
+.. automodule:: xgi.convert.graph
+   
+   .. rubric:: Functions
+   
+   .. autofunction:: to_graph
