+++ docs/source/api/convert/xgi.convert.higher_order_network.rst
@@ -1,0 +1,12 @@
+xgi.convert.higher_order_network
+================================
+
+.. currentmodule:: xgi.convert.higher_order_network
+
+.. automodule:: xgi.convert.higher_order_network
+   
+   .. rubric:: Functions
+   
+   .. autofunction:: convert_to_hypergraph
+   .. autofunction:: convert_to_dihypergraph
+   .. autofunction:: convert_to_simplicial_complex
