+++ docs/source/api/convert/xgi.convert.hyperedges.rst
@@ -1,0 +1,13 @@
+xgi.convert.hyperedges
+======================
+
+.. currentmodule:: xgi.convert.hyperedges
+
+.. automodule:: xgi.convert.hyperedges
+   
+   .. rubric:: Functions
+   
+   .. autofunction:: from_hyperedge_dict
+   .. autofunction:: to_hyperedge_dict
+   .. autofunction:: from_hyperedge_list
+   .. autofunction:: to_hyperedge_list
