+++ docs/source/api/convert/xgi.convert.hypergraph_dict.rst
@@ -1,0 +1,10 @@
+xgi.convert.hypergraph_dict
+===========================
+
+.. currentmodule:: xgi.convert.hypergraph_dict
+
+.. automodule:: xgi.convert.hypergraph_dict
+   
+   .. rubric:: Functions
+   
+   .. autofunction:: dict_to_hypergraph
