+++ docs/source/api/convert/xgi.convert.incidence.rst
@@ -1,0 +1,11 @@
+xgi.convert.incidence
+=====================
+
+.. currentmodule:: xgi.convert.incidence
+
+.. automodule:: xgi.convert.incidence
+   
+   .. rubric:: Functions
+   
+   .. autofunction:: from_incidence_matrix
+   .. autofunction:: to_incidence_matrix
