+++ docs/source/api/convert/xgi.convert.line_graph.rst
@@ -1,0 +1,10 @@
+xgi.convert.line_graph
+======================
+
+.. currentmodule:: xgi.convert.line_graph
+
+.. automodule:: xgi.convert.line_graph
+   
+   .. rubric:: Functions
+   
+   .. autofunction:: to_line_graph
