+++ docs/source/api/convert/xgi.convert.pandas.rst
@@ -1,0 +1,11 @@
+xgi.convert.pandas
+==================
+
+.. currentmodule:: xgi.convert.pandas
+
+.. automodule:: xgi.convert.pandas
+   
+   .. rubric:: Functions
+   
+   .. autofunction:: from_bipartite_pandas_dataframe
+   .. autofunction:: to_bipartite_pandas_dataframe
