+++ docs/source/api/convert/xgi.convert.simplex.rst
@@ -1,0 +1,11 @@
+xgi.convert.simplex
+===================
+
+.. currentmodule:: xgi.convert.simplex
+
+.. automodule:: xgi.convert.simplex
+   
+   .. rubric:: Functions
+   
+   .. autofunction:: from_max_simplices
+   .. autofunction:: from_simplex_dict
