+++ xgi/convert/__init__.py
@@ -1,0 +1,22 @@
+from . import (
+    bipartite_edges,
+    bipartite_graph,
+    graph,
+    higher_order_network,
+    hyperedges,
+    hypergraph_dict,
+    incidence,
+    line_graph,
+    pandas,
+    simplex,
+)
+from .bipartite_edges import *
+from .bipartite_graph import *
+from .graph import *
+from .higher_order_network import *
+from .hyperedges import *
+from .hypergraph_dict import *
+from .incidence import *
+from .line_graph import *
+from .pandas import *
+from .simplex import *
